feat: Introduced a Chongqing Assassination challenge to unlock the suit.
For Peacock only, for those who want to earn the outfit. Challenge unlock is the default behavior. You don't have to, unlocking it from the start is available as a mod option as well, which is the old behavior.

Complete the Hands Dirtied challenge in End Of An Era by eliminating Hush and Imogen Royce with a Fiber Wire to unlock the suit.